Feat: useSwipeVector - Code splitting - Publish beta action
