+++ .github/workflows/npm-publish.yml
@@ -14,7 +14,7 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
-    if: github.event_name == 'push' && !(github.ref == 'refs/heads/master' && contains(github.event.head_commit.message, '[publish]'))
+    if: github.event_name == 'push' && !(github.ref == 'refs/heads/master' && contains(github.event.head_commit.message, '[publish'))
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
@@ -38,5 +38,21 @@ jobs:
         run: npm ci
       - name: Publish
         run: npm publish
+        env:
+          NODE_AUTH_TOKEN: ${{secrets.npm_token}}
+
+  publish-npm-beta:
+    runs-on: ubuntu-latest
+    if: github.ref == 'refs/heads/master' && github.event_name == 'push' && contains(github.event.head_commit.message, '[publish-beta]')
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v2
+        with:
+          node-version: 12
+          registry-url: https://registry.npmjs.org/
+      - name: Install
+        run: npm ci
+      - name: Publish
+        run: npm publish --tag beta
         env:
           NODE_AUTH_TOKEN: ${{secrets.npm_token}}
