+++ package.json
@@ -1,9 +1,10 @@
 {
   "name": "use-swipe-hook",
-  "version": "1.2.2",
+  "version": "2.0.0-beta1",
   "description": "A simple and easy to use tiny library that provides useSwipe hook to use with React that enables swipe gestures for touch screens",
-  "main": "lib/index.js",
-  "browser": "lib/index.js",
+  "main": "lib/umd/index.js",
+  "browser": "lib/umd/index.js",
+  "module": "lib/esm/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "rollup -c",
