+++ src/helper.ts
@@ -1,0 +1,7 @@
+export function square(x: number) {
+  return x * x;
+}
+
+export function radToDeg(x: number) {
+  return x * 360 / Math.PI;
+}
