+++ src/index.ts
@@ -1,2 +1,3 @@
-export { default as useSwipe } from './useSwipe';
-export * from './useSwipe';
+export { default as useSwipe, SWIPE_DIRECTION } from './useSwipe';
+export { default as useSwipePosition } from './useSwipePosition';
+export { default as useSwipeVector } from './useSwipeVector';
