+++ src/useSwipe.ts
@@ -1,4 +1,6 @@
-import { RefObject, useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
+import { UseSwipeOptions } from './types';
+import useSwipePosition from './useSwipePosition';
 
 // SWIPE_DIRECTION can be imported & used for comparison
 export enum SWIPE_DIRECTION {
@@ -8,19 +10,8 @@ export enum SWIPE_DIRECTION {
   DOWN = "down"
 }
 
-interface UseSwipeOptions {
-    // ref of the container where you want to attach swipe event
-    ref: RefObject<HTMLElement>;
-    // (optional) no of pixels to move your finger to trigger a swipe event. 
-    // Larger this value means less sensitivity. Default value is 5 (5px)
-    thresholdPX?: number; 
-};
-
 const useSwipe = ({ ref, thresholdPX = 5 }: UseSwipeOptions) => {
-  const [x1, setX1] = useState(0);
-  const [y1, setY1] = useState(0);
-  const [x2, setX2] = useState(0);
-  const [y2, setY2] = useState(0);
+  const { x1, y1, x2, y2 } = useSwipePosition({ ref });
   const [direction, setDirection] = useState<SWIPE_DIRECTION | null>(null);
 
   useEffect(() => {
@@ -32,59 +23,7 @@ const useSwipe = ({ ref, thresholdPX = 5 }: UseSwipeOptions) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [y2]);
 
-  useEffect(() => {
-    const currentElement = ref.current;
-
-    if (currentElement) {
-      if (isTouchDevice()) {
-        currentElement.addEventListener('touchstart', handleTouchStart);
-        currentElement.addEventListener('touchend', handleTouchEnd);
-      }
-      else {
-        currentElement.addEventListener('mousedown', handleMouseDown);
-        currentElement.addEventListener('mouseup', handleMouseUp);
-      }
-    }
-
-    return () => {
-      if (currentElement) {
-        if (isTouchDevice()) {
-          currentElement.removeEventListener('touchstart', handleTouchStart);
-          currentElement.removeEventListener('touchend', handleTouchEnd);
-        }
-        else {
-          currentElement.removeEventListener('mousedown', handleMouseDown);
-          currentElement.removeEventListener('mouseup', handleMouseUp);
-        }
-      }
-    }
-  }, [ref]);
-
-  function handleTouchStart(event: TouchEvent) {
-    setX1(event.changedTouches[0].clientX);
-    setY1(event.changedTouches[0].clientY);
-  }
-
-  function handleTouchEnd(event: TouchEvent) {
-    if (event.changedTouches && event.changedTouches.length > 0) {
-      setX2(event.changedTouches[0].clientX);
-      setY2(event.changedTouches[0].clientY);
-    }
-  }
-
-  function handleMouseDown(event: MouseEvent) {
-    setX1(event.clientX);
-    setY1(event.clientY);
-  }
-
-  function handleMouseUp(event: MouseEvent) {
-    setX2(event.clientX);
-    setY2(event.clientY);
-  }
-
   return direction;
 };
-
-const isTouchDevice = () => ('ontouchstart' in window);
 
 export default useSwipe;
